Create Util.key to return key from an event
Instead of always trying to find which key is associated to an event for all browser, I created a helper in Util to return this key.

I follow what is already done by jQuery.

It'll ease the transition when the `.key` will be compatible with all browsers (https://developer.mozilla.org/en-US/docs/Web/API/KeyboardEvent/key)